# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PropostaPrazo": {
        "description": "Proposta com prazos vencendo na janela consultada (visão estruturada).",
        "properties": {
          "prazos_na_janela": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Prazos Na Janela",
            "type": "array"
          },
          "proposta": {
            "$ref": "#/components/schemas/PropostaRead"
          }
        },
        "required": [
          "proposta",
          "prazos_na_janela"
        ],
        "title": "PropostaPrazo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "API pública v1 — propostas/repasses do governo (cache-first, RLS por usuário).",
    "title": "Hub Capture API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/proposals": {
      "get": {
        "operationId": "listar_propostas_api_v1_proposals_get",
        "parameters": [
          {
            "description": "código IBGE (7 dígitos)",
            "in": "query",
            "name": "municipio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "código IBGE (7 dígitos)",
              "title": "Municipio"
            }
          },
          {
            "in": "query",
            "name": "fonte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fonte"
            }
          },
          {
            "description": "área de interesse (saude, educacao…)",
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "área de interesse (saude, educacao…)",
              "title": "Area"
            }
          },
          {
            "in": "query",
            "name": "situacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Situacao"
            }
          },
          {
            "description": "tipo de instrumento",
            "in": "query",
            "name": "modalidade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tipo de instrumento",
              "title": "Modalidade"
            }
          },
          {
            "description": "órgão/ministério concedente",
            "in": "query",
            "name": "orgao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "órgão/ministério concedente",
              "title": "Orgao"
            }
          },
          {
            "description": "municipal | estadual_df | consorcio | empresa_publica | osc",
            "in": "query",
            "name": "natureza_juridica",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "municipal | estadual_df | consorcio | empresa_publica | osc",
              "title": "Natureza Juridica"
            }
          },
          {
            "description": "tipo de transferência",
            "in": "query",
            "name": "qualificacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tipo de transferência",
              "title": "Qualificacao"
            }
          },
          {
            "in": "query",
            "name": "ano",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ano"
            }
          },
          {
            "description": "busca por programa, órgão ou código",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "busca por programa, órgão ou código",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "valor_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valor Min"
            }
          },
          {
            "in": "query",
            "name": "valor_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valor Max"
            }
          },
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(cadastrada|disponivel)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "description": "mais recentes · prazo (próximo/distante) · nome A-Z · órgão A-Z · valor",
            "in": "query",
            "name": "ordenar",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(recentes|prazo|prazo_distante|nome|orgao|valor)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "mais recentes · prazo (próximo/distante) · nome A-Z · órgão A-Z · valor",
              "title": "Ordenar"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropostaRead"
                  },
                  "title": "Response Listar Propostas Api V1 Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Propostas",
        "tags": [
          "propostas"
        ]
      }
    },
    "/api/v1/proposals/deadlines": {
      "get": {
        "description": "Propostas com prazo vencendo na janela — 'quais vencem este mês?'.",
        "operationId": "propostas_por_prazo_api_v1_proposals_deadlines_get",
        "parameters": [
          {
            "in": "query",
            "name": "dias",
            "required": false,
            "schema": {
              "default": 30,
              "maximum": 365,
              "minimum": 1,
              "title": "Dias",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PropostaPrazo"
                  },
                  "title": "Response Propostas Por Prazo Api V1 Proposals Deadlines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Propostas Por Prazo",
        "tags": [
          "propostas"
        ]
      }
    },
    "/api/v1/proposals/facets": {
      "get": {
        "description": "Opções de cada filtro com contagem — alimenta os dropdowns da tela.",
        "operationId": "facetas_propostas_api_v1_proposals_facets_get",
        "parameters": [
          {
            "description": "código IBGE (7 dígitos)",
            "in": "query",
            "name": "municipio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "código IBGE (7 dígitos)",
              "title": "Municipio"
            }
          },
          {
            "in": "query",
            "name": "fonte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fonte"
            }
          },
          {
            "description": "área de interesse (saude, educacao…)",
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "área de interesse (saude, educacao…)",
              "title": "Area"
            }
          },
          {
            "in": "query",
            "name": "situacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Situacao"
            }
          },
          {
            "description": "tipo de instrumento",
            "in": "query",
            "name": "modalidade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tipo de instrumento",
              "title": "Modalidade"
            }
          },
          {
            "description": "órgão/ministério concedente",
            "in": "query",
            "name": "orgao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "órgão/ministério concedente",
              "title": "Orgao"
            }
          },
          {
            "description": "municipal | estadual_df | consorcio | empresa_publica | osc",
            "in": "query",
            "name": "natureza_juridica",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "municipal | estadual_df | consorcio | empresa_publica | osc",
              "title": "Natureza Juridica"
            }
          },
          {
            "description": "tipo de transferência",
            "in": "query",
            "name": "qualificacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tipo de transferência",
              "title": "Qualificacao"
            }
          },
          {
            "in": "query",
            "name": "ano",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ano"
            }
          },
          {
            "description": "busca por programa, órgão ou código",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "busca por programa, órgão ou código",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "valor_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valor Min"
            }
          },
          {
            "in": "query",
            "name": "valor_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valor Max"
            }
          },
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(cadastrada|disponivel)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropostasFacetas"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Facetas Propostas",
        "tags": [
          "propostas"
        ]
      }
    },
...
    "/api/v1/proposals/report.csv": {
      "get": {
        "description": "Baixar relatório: o mesmo recorte da tela, em CSV (';', abre no Excel).",
        "operationId": "relatorio_propostas_api_v1_proposals_report_csv_get",
        "parameters": [
          {
            "description": "código IBGE (7 dígitos)",
            "in": "query",
            "name": "municipio",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "código IBGE (7 dígitos)",
              "title": "Municipio"
            }
          },
          {
            "in": "query",
            "name": "fonte",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fonte"
            }
          },
          {
            "description": "área de interesse (saude, educacao…)",
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "área de interesse (saude, educacao…)",
              "title": "Area"
            }
          },
          {
            "in": "query",
            "name": "situacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Situacao"
            }
          },
          {
            "description": "tipo de instrumento",
            "in": "query",
            "name": "modalidade",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tipo de instrumento",
              "title": "Modalidade"
            }
          },
          {
            "description": "órgão/ministério concedente",
            "in": "query",
            "name": "orgao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "órgão/ministério concedente",
              "title": "Orgao"
            }
          },
          {
            "description": "municipal | estadual_df | consorcio | empresa_publica | osc",
            "in": "query",
            "name": "natureza_juridica",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "municipal | estadual_df | consorcio | empresa_publica | osc",
              "title": "Natureza Juridica"
            }
          },
          {
            "description": "tipo de transferência",
            "in": "query",
            "name": "qualificacao",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tipo de transferência",
              "title": "Qualificacao"
            }
          },
          {
            "in": "query",
            "name": "ano",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ano"
            }
          },
          {
            "description": "busca por programa, órgão ou código",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "busca por programa, órgão ou código",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "valor_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valor Min"
            }
          },
          {
            "in": "query",
            "name": "valor_max",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0.0,
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Valor Max"
            }
          },
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(cadastrada|disponivel)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "description": "mais recentes · prazo (próximo/distante) · nome A-Z · órgão A-Z · valor",
            "in": "query",
            "name": "ordenar",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(recentes|prazo|prazo_distante|nome|orgao|valor)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "mais recentes · prazo (próximo/distante) · nome A-Z · órgão A-Z · valor",
              "title": "Ordenar"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Relatorio Propostas",
        "tags": [
          "propostas"
        ]
      }
    },
...
    "/api/v1/proposals/{proposta_id}": {
      "get": {
        "operationId": "obter_proposta_api_v1_proposals__proposta_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposta Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropostaRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Obter Proposta",
        "tags": [
          "propostas"
        ]
      }
    },
    "/api/v1/proposals/{proposta_id}/pdf": {
      "get": {
        "operationId": "exportar_pdf_api_v1_proposals__proposta_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "proposta_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Proposta Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Exportar Pdf",
        "tags": [
          "propostas"
        ]
      }
    },
...
  }
}